requirements.txt
Django==1.11.29
future
captcha
django-compressor
django-statici18n
django-post_office==3.3.0
django-webpack_loader
gunicorn
pymysql
django-picklefield
openpyxl
qrcode
django-formtools
django-simple-captcha
djangorestframework
python-dateutil
requests
pillow
pyjwt
python-ldap
pysaml2
aliyun-python-sdk-core-v3==2.13.11
urllib3>=1.25.8
requests_oauthlib
